+++ remote/sdk/bcl/inc/bc_radio.h
@@ -34,6 +34,6 @@ bool bc_radio_pub_luminosity(uint8_t i2c, float *lux);
 
 bool bc_radio_pub_barometer(uint8_t i2c, float *pascal, float *meter);
 
-bool bc_radio_pub_co2(uint8_t i2c, int16_t *concentration);
+bool bc_radio_pub_co2(uint8_t i2c, float *concentration);
 
 #endif // _BC_RADIO_H
